pushing

+++ src/classes/ClassConditionalStbleTgtMarkovianPostMeanDiffTrainer.py
@@ -655,7 +655,7 @@ class ConditionalStbleTgtMarkovianPostMeanDiffTrainer(nn.Module):
                                         diff=config.diffusion,
                                         initial_state=config.initState)
             data = data.astype(float)
-            Xs = data.reshape((-1, config.ndims), order="C", dtype=torch.float32)
+            Xs = data.reshape((-1, config.ndims), order="C")
             final_vec_mu_hats = experiment_MLP_DDims_drifts(config=config, Xs=Xs, good=self.score_network.module,
                                                             onlyGauss=False)
         else:
